Hotfix: Integer division shouldn't happen, division should return decimals now

+++ src/main/java/hello/SampleController.java
@@ -61,7 +61,7 @@ public class SampleController {
     JSONObject operands = new JSONObject(jsonString);
     long num1 = operands.getInt("num1");
     long num2 = operands.getInt("num2");
-    long result = num1 / num2;
+    double result = (double) num1 / num2;
 
     // Sanity Check
     System.out.println("num1: " + new Long(num1).toString() + "num2: " + new Long(num2).toString());
